move fill_binop
